Rename libvirt and vcenter directory because of collision
